+++ src/components/SQForm/SQFormTextField.js
@@ -16,7 +16,9 @@ function SQFormTextField({
   onBlur,
   onChange,
   startAdornment,
-  endAdornment
+  endAdornment,
+  type = 'text',
+  muiFieldProps = {}
 }) {
   const {
     formikField: {field},
@@ -47,7 +49,7 @@ function SQFormTextField({
         }}
         FormHelperTextProps={{error: isFieldError}}
         name={name}
-        type="text"
+        type={type}
         label={label}
         helperText={HelperTextComponent}
         placeholder={placeholder}
@@ -55,6 +57,7 @@ function SQFormTextField({
         onBlur={handleBlur}
         required={isRequired}
         value={field.value}
+        {...muiFieldProps}
       />
     </Grid>
   );
@@ -80,7 +83,11 @@ SQFormTextField.propTypes = {
   /** Adornment that appears at the start of the input */
   startAdornment: PropTypes.node,
   /** Adornment that appears at the end of the input */
-  endAdornment: PropTypes.node
+  endAdornment: PropTypes.node,
+  /** Defines the input type for the text field. Must be a valid HTML5 input type */
+  type: PropTypes.string,
+  /** Any valid prop for material ui text input child component - https://material-ui.com/api/text-field/#props */
+  muiFieldProps: PropTypes.object
 };
 
 export default SQFormTextField;